Chore: Move app.js to server folder
