+++ server/tests/routes/MenuRoutes.js/testGetMenuForSpecificDateRoute.js
@@ -1,6 +1,6 @@
 import chai from 'chai';
 import chaiHttp from 'chai-http';
-import App from "../../../../app";
+import App from "../../../app";
 
 chai.use(chaiHttp);
 chai.should();
